OAI-64: Move fhir sub-module to main package
