+++ claim_ai_quality/ai_evaluation/integrated_evaluation_organizer.py
@@ -7,7 +7,7 @@ from claim_ai.evaluation.converters.r4_fhir_resources.fhir_response_builders.bas
 from claim_ai.rest_api.claim_evaluation.claim_bundle_evaluation_manager import ClaimBundleEvaluationManager
 from claim_ai_quality.ai_evaluation.mutation_evaluation import EvaluationFromMutation
 from claim_ai_quality.apps import ClaimAiQualityConfig
-from claim_ai_quality.communication_interface.fhir._claim_response_converter import ClaimResponseConverter
+from claim_ai_quality.fhir._claim_response_converter import ClaimResponseConverter
 from claim_ai_quality.models import ClaimBundleEvaluationResult, BundleClaimEvaluationResult
 from claim_ai_quality.utils import reset_sent_but_not_evaluated_claims, add_json_ext_to_all_submitted_claims, \
     get_eligible_claims_bundle_iterator
