+++ claim_ai_quality/fhir/fhirConverter.py
@@ -8,7 +8,7 @@ from api_fhir_r4.converters import ReferenceConverterMixin
 from api_fhir_r4.models import BundleType
 from api_fhir_r4.serializers import ClaimSerializer
 from claim.models import Claim, ClaimItem, ClaimService
-from claim_ai_quality.communication_interface.fhir._claim_response_converter import ClaimResponseConverter
+from claim_ai_quality.fhir._claim_response_converter import ClaimResponseConverter
 
 logger = logging.getLogger(__name__)
